+++ panoptica/utils/__init__.py
@@ -40,10 +40,10 @@ from panoptica.utils.serialization import (
 )
 from panoptica.utils.file_backend import (
     FileBackend,
-    get_backend,
     FileType,
     supported_file_types,
     derive_file_type,
 )
 from panoptica.utils.file_backend_jsonl import JSONLBackend
 from panoptica.utils.file_backend_tsv import TSVBackend
+from panoptica.utils.file_backend_registry import get_backend
